+++ CHANGELOG.md
@@ -53,12 +53,13 @@ Before releasing:
 - Added implementations of `Mul<i64>` and `Div<i64>` for `Position`, allowing
   for opaque scaling (#230)
 - Added panic hook support comparable to the Rust standard library through `vexide::panic::set_hook` and `vexide::panic::take_hook` (#234)
+- Added support for legacy ADI servos through the `AdiServo` API. (#241)
 - Added support for the V5 AI Vision Sensor (#58)
 - Added FOV constants to the Vision Sensor (#58)
 - Added missing `Send` and `Sync` `impl`s for RwLock. (#239)
 - Added the `Proportional` font family and support for fractional font scaling. (#248) (**Breaking Change**)
 - Added `AdiDigitalOut::with_initial_state` to set the initial state of a digital output while creating it (#246)
-- Added support for the legacy Yaw Rate Gyroscope through the `AdiGyro` struct. (#236)
+- Added support for the legacy Yaw Rate Gyroscope through the `AdiGyroscope` struct. (#236)
 
 ### Fixed